+++ CHANGELOG.md
@@ -2,12 +2,24 @@
 
 <!-- ## [Unreleased] -->
 
+## [3.0.0] - 2023-07-26
+
+- Unregister hook handlers on unmount
+- Autobind hook handlers to plugin instance
+- Update for compatbility with swup 4
+
 ## [2.0.0] - 2023-01
 
 - Allow plugins to specify swup version requirements
 - Switch to microbundle for bundling
 - Provide a simple CLI for plugins to bundle and lint files
 
-[Unreleased]: https://github.com/swup/plugin/compare/2.0.0...HEAD
+## [1.0.0] - 2019-05-20
 
+- Initial release
+
+[Unreleased]: https://github.com/swup/plugin/compare/3.0.0...HEAD
+
+[3.0.0]: https://github.com/swup/plugin/releases/tag/3.0.0
 [2.0.0]: https://github.com/swup/plugin/releases/tag/2.0.0
+[1.0.0]: https://github.com/swup/plugin/releases/tag/1.0.0